+++ product/juno/module/juno_rom/src/juno_debug_rom.c
@@ -213,31 +213,31 @@ int juno_debug_rom_init(const struct mod_juno_ppu_rom_api *rom_ppu_api)
             SCP_CONFIG_DEBUG_CONTROL_CDBGPWRUPACK) != 0);
     }
 
-    fwk_interrupt_clear_pending(CDBG_PWR_UP_REQ_IRQ);
-    fwk_interrupt_clear_pending(CDBG_RST_REQ_IRQ);
-    fwk_interrupt_clear_pending(CSYS_PWR_UP_REQ_IRQ);
+    fwk_interrupt_clear_pending((unsigned int)CDBG_PWR_UP_REQ_IRQ);
+    fwk_interrupt_clear_pending((unsigned int)CDBG_RST_REQ_IRQ);
+    fwk_interrupt_clear_pending((unsigned int)CSYS_PWR_UP_REQ_IRQ);
 
-    status = fwk_interrupt_set_isr(CDBG_PWR_UP_REQ_IRQ,
-        juno_debug_cdbg_pwr_up_req_isr);
+    status = fwk_interrupt_set_isr(
+        (unsigned int)CDBG_PWR_UP_REQ_IRQ, juno_debug_cdbg_pwr_up_req_isr);
     if (status != FWK_SUCCESS) {
         return status;
     }
 
-    status = fwk_interrupt_set_isr(CSYS_PWR_UP_REQ_IRQ,
-        juno_debug_csys_pwr_up_req_isr);
+    status = fwk_interrupt_set_isr(
+        (unsigned int)CSYS_PWR_UP_REQ_IRQ, juno_debug_csys_pwr_up_req_isr);
     if (status != FWK_SUCCESS) {
         return status;
     }
 
-    status = fwk_interrupt_set_isr(CDBG_RST_REQ_IRQ,
-        juno_debug_cdbg_rst_req_isr);
+    status = fwk_interrupt_set_isr(
+        (unsigned int)CDBG_RST_REQ_IRQ, juno_debug_cdbg_rst_req_isr);
     if (status != FWK_SUCCESS) {
         return status;
     }
 
-    fwk_interrupt_enable(CDBG_PWR_UP_REQ_IRQ);
-    fwk_interrupt_enable(CDBG_RST_REQ_IRQ);
-    fwk_interrupt_enable(CSYS_PWR_UP_REQ_IRQ);
+    fwk_interrupt_enable((unsigned int)CDBG_PWR_UP_REQ_IRQ);
+    fwk_interrupt_enable((unsigned int)CDBG_RST_REQ_IRQ);
+    fwk_interrupt_enable((unsigned int)CSYS_PWR_UP_REQ_IRQ);
 
     return FWK_SUCCESS;
 }
