+++ product/juno/scp_ramfw/config_scmi_system_power.c
@@ -16,7 +16,8 @@ struct fwk_module_config config_scmi_system_power = {
     .data =
         &(struct mod_scmi_system_power_config){
             .system_view = MOD_SCMI_SYSTEM_VIEW_FULL,
-            .system_suspend_state = MOD_SYSTEM_POWER_POWER_STATE_SLEEP0,
+            .system_suspend_state =
+                (unsigned int)MOD_SYSTEM_POWER_POWER_STATE_SLEEP0,
             .alarm_id = FWK_ID_SUB_ELEMENT_INIT(
                 FWK_MODULE_IDX_TIMER,
                 JUNO_ALARM_ELEMENT_IDX,
